More math refactoring, remove dependency on GLM for quaternions. lerp no longer clamps.
